v2/requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress & Logging
tqdm>=4.65.0

# Utilities
pyyaml>=6.0